Verified Dockerfile, travis and updated classpaths
