+++ sink.properties
@@ -2,7 +2,7 @@
 name=kafka-connector-mongodb-sink
 
 # Kafka connector configuration
-connector.class=org.radarcns.mongodb.MongoDbSinkConnector
+connector.class=org.radarcns.connect.mongodb.MongoDbSinkConnector
 tasks.max=1
 
 # Topics that will be consumed
@@ -21,4 +21,4 @@ mongo.database=
 #mongo.collection.format={$topic}
 
 # Factory class to do the actual record conversion
-record.converter.class=org.radarcns.serialization.RecordConverterFactory
+#record.converter.class=org.radarcns.connect.mongodb.serialization.RecordConverterFactory
