+++ src/main/docker/launch
@@ -38,7 +38,7 @@ echo "===> Wait for infrastructure ..."
 kafka-wait
 radar_check=$?
 if [ "$radar_check" -ne 0 ]; then
-    exit $radar_check
+    exit ${radar_check}
 fi
 
 echo "===> Launching ${COMPONENT} ... "
@@ -46,4 +46,4 @@ echo "===> Launching ${COMPONENT} ... "
 # And this also makes sure that the CLASSPATH does not start with ":/etc/..."
 # because this causes the plugin scanner to scan the entire disk.
 export CLASSPATH="/etc/kafka-connect/jars/*"
-exec connect-standalone /etc/"${COMPONENT}"/"${COMPONENT}".properties /etc/"${COMPONENT}"/sink.properties
+exec connect-standalone /etc/kafka-connect/kafka-connect.properties /etc/kafka-connect/sink.properties
