--- src/main/docker/sink-mongodb.properties
@@ -1,25 +1,0 @@
-# Kafka consumer configuration
-name=radar-connector-mongodb-sink
-
-# Kafka connector configuration
-connector.class=org.radarcns.mongodb.MongoDbSinkConnector
-tasks.max=1
-
-# Topics that will be consumed
-topics=android_empatica_e4_battery_level,android_empatica_e4_battery_level_output
-
-# MongoDB server
-mongo.host=mongo
-mongo.port=27017
-
-# MongoDB configuration
-#mongo.username=
-#mongo.password=
-mongo.database=mydbase
-
-# Collection name for putting data into the MongoDB database. The {$topic} token will be replaced
-# by the Kafka topic name.
-#mongo.collection.format={$topic}
-
-# Factory class to do the actual record conversion
-record.converter.class=org.radarcns.sink.mongodb.RecordConverterFactoryRadar
