+++ src/main/java/org/radarcns/connect/mongodb/serialization/RecordConverterFactory.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.radarcns.serialization;
+package org.radarcns.connect.mongodb.serialization;
 
 import java.util.Collections;
 import java.util.HashMap;
