+++ src/test/java/org/radarcns/connect/mongodb/MongoDbSinkConnectorTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.radarcns.mongodb;
+package org.radarcns.connect.mongodb;
 
 import static org.apache.kafka.connect.sink.SinkConnector.TOPICS_CONFIG;
 import static org.hamcrest.Matchers.contains;
@@ -22,13 +22,13 @@ import static org.hamcrest.Matchers.hasSize;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertThat;
-import static org.radarcns.mongodb.MongoDbSinkConnector.BUFFER_CAPACITY;
-import static org.radarcns.mongodb.MongoDbSinkConnector.COLLECTION_FORMAT;
-import static org.radarcns.mongodb.MongoDbSinkConnector.MONGO_DATABASE;
-import static org.radarcns.mongodb.MongoDbSinkConnector.MONGO_HOST;
-import static org.radarcns.mongodb.MongoDbSinkConnector.MONGO_PORT;
-import static org.radarcns.mongodb.MongoDbSinkConnector.MONGO_PORT_DEFAULT;
-import static org.radarcns.mongodb.MongoDbSinkConnector.MONGO_USERNAME;
+import static org.radarcns.connect.mongodb.MongoDbSinkConnector.BUFFER_CAPACITY;
+import static org.radarcns.connect.mongodb.MongoDbSinkConnector.COLLECTION_FORMAT;
+import static org.radarcns.connect.mongodb.MongoDbSinkConnector.MONGO_DATABASE;
+import static org.radarcns.connect.mongodb.MongoDbSinkConnector.MONGO_HOST;
+import static org.radarcns.connect.mongodb.MongoDbSinkConnector.MONGO_PORT;
+import static org.radarcns.connect.mongodb.MongoDbSinkConnector.MONGO_PORT_DEFAULT;
+import static org.radarcns.connect.mongodb.MongoDbSinkConnector.MONGO_USERNAME;
 
 import java.util.Collection;
 import java.util.HashMap;
