+++ src/test/java/org/radarcns/connect/mongodb/MongoDbSinkTaskTest.java
@@ -14,18 +14,18 @@
  * limitations under the License.
  */
 
-package org.radarcns.mongodb;
+package org.radarcns.connect.mongodb;
 
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
-import static org.radarcns.mongodb.MongoDbSinkConnector.BUFFER_CAPACITY;
-import static org.radarcns.mongodb.MongoDbSinkConnector.MONGO_DATABASE;
-import static org.radarcns.mongodb.MongoDbSinkConnector.MONGO_HOST;
-import static org.radarcns.mongodb.MongoDbSinkConnector.RECORD_CONVERTER;
-import static org.radarcns.mongodb.MongoDbSinkConnector.TOPICS_CONFIG;
+import static org.radarcns.connect.mongodb.MongoDbSinkConnector.BUFFER_CAPACITY;
+import static org.radarcns.connect.mongodb.MongoDbSinkConnector.MONGO_DATABASE;
+import static org.radarcns.connect.mongodb.MongoDbSinkConnector.MONGO_HOST;
+import static org.radarcns.connect.mongodb.MongoDbSinkConnector.RECORD_CONVERTER;
+import static org.radarcns.connect.mongodb.MongoDbSinkConnector.TOPICS_CONFIG;
 
 import java.util.Collections;
 import java.util.HashMap;
@@ -36,7 +36,7 @@ import org.apache.kafka.connect.sink.SinkRecord;
 import org.junit.Test;
 import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
-import org.radarcns.serialization.RecordConverterFactory;
+import org.radarcns.connect.mongodb.serialization.RecordConverterFactory;
 
 public class MongoDbSinkTaskTest {
