+++ src/test/java/org/radarcns/connect/util/ValidClassTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.radarcns.util;
+package org.radarcns.connect.util;
 
 import static org.junit.Assert.assertNotNull;
