--- standalone.properties
@@ -1,32 +1,0 @@
-# Sample configuration for a standalone Kafka Connect worker that uses Avro serialization and
-# integrates the the Schema Registry. This sample configuration assumes a local installation of
-# Confluent Platform with all services running on their default ports.
-
-# Bootstrap Kafka servers. If multiple servers are specified, they should be comma-separated.
-bootstrap.servers=localhost:9092
-consumer.max.poll.records=8000
-consumer.auto.offset.reset=earliest
-
-# The converters specify the format of data in Kafka and how to translate it into Connect data.
-# Every Connect user will need to configure these based on the format they want their data in
-# when loaded from or stored into Kafka
-key.converter=io.confluent.connect.avro.AvroConverter
-key.converter.schema.registry.url=http://localhost:8081
-value.converter=io.confluent.connect.avro.AvroConverter
-value.converter.schema.registry.url=http://localhost:8081
-
-# The internal converter used for offsets and config data is configurable and must be specified,
-# but most users will always want to use the built-in default. Offset and config data is never
-# visible outside of Connect in this format.
-internal.key.converter=org.apache.kafka.connect.json.JsonConverter
-internal.value.converter=org.apache.kafka.connect.json.JsonConverter
-internal.key.converter.schemas.enable=false
-internal.value.converter.schemas.enable=false
-
-# Local storage file for offset data
-offset.storage.file.filename=/tmp/connect_mongo.offsets
-
-# Confuent Control Center Integration -- uncomment these lines to enable Kafka client interceptors
-# that will report audit data that can be displayed and analyzed in Confluent Control Center
-# producer.interceptor.classes=io.confluent.monitoring.clients.interceptor.MonitoringProducerInterceptor
-# consumer.interceptor.classes=io.confluent.monitoring.clients.interceptor.MonitoringConsumerInterceptor
